+++ DojoManagmentSystem/DojoManagmentSystem/ViewModels/AttendanceViewModel.cs
@@ -18,9 +18,9 @@ namespace DojoManagmentSystem.ViewModels
 
         public string FullName { get; set; }
 
-        public int MemberId { get; set; }
+        public long MemberId { get; set; }
 
-        public int ClassSessionId { get; set; }
+        public long ClassSessionId { get; set; }
 
         public virtual ClassSession ClassSession { get; set; }
     }
